feat: LambdaFunctions added
 - serverless
 - handler.js
 - create-layer
 - scraper caller
 - users model refactored
 - aws-sdk added
