+++ README.md
@@ -6,6 +6,24 @@ Import your Uninorte schedule to Google Calendar.
 ```
   docker-compose up
 ```
+#### Serverless
+* Required tools:
+```
+  npm install -g serverless
+  brew install brotli
+```
+* Compress required libaries
+```
+  cd src/lambdaFunctions/scraper
+  nvm use 8.11
+  ./create-layer
+```
+* Upload mihorarioun-scraper-layer.zip layer
+* Deploy
+```
+  serverless config credentials --provider aws --key KEY --secret SECRET_KEY
+  serverless deploy
+```
 ### ui
 ```
   cd ui && ng serve
