+++ server/package.json
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/sjdonado/mihorario#readme",
   "dependencies": {
+    "aws-sdk": "^2.506.0",
     "body-parser": "^1.18.3",
     "cors": "^2.8.5",
     "express": "^4.16.4",
