+++ server/src/lambdaFunctions/scraper/create-layer
@@ -1,0 +1,14 @@
+#!/bin/bash
+git clone --depth=1 https://github.com/alixaxel/chrome-aws-lambda.git && \
+cd chrome-aws-lambda && \
+brotli --decompress --rm bin/chromium-*.br && \
+npm pack && \
+mkdir -p nodejs/node_modules/chrome-aws-lambda/ && \
+tar --directory nodejs/node_modules/chrome-aws-lambda/ --extract --file chrome-aws-lambda-*.tgz --strip-components=1 && \
+rm chrome-aws-lambda-*.tgz && \
+npm install moment  --no-bin-links --no-optional --no-package-lock --no-save --no-shrinkwrap --prefix nodejs/ && \
+npm install puppeteer-core --no-bin-links --no-optional --no-package-lock --no-save --no-shrinkwrap --prefix nodejs/ && \
+zip -9 --filesync --move --recurse-paths ./mihorarioun-scraper-layer.zip nodejs/ && \
+mv ./mihorarioun-scraper-layer.zip ../mihorarioun-scraper-layer.zip && \
+cd ../ && \
+sudo rm -r chrome-aws-lambda
