+++ ui/src/app/components/help/help.component.scss
@@ -3,10 +3,6 @@ mat-card {
   max-width: 80%;
 }
 
-.app-name {
-  font-size: 20px;
-}
-
 .info {
   width: 100%;
   text-align: center;
