+++ ui/src/app/components/index/index.component.scss
@@ -1,5 +1,6 @@
 .app-name {
   font-size: 24px;
+  text-align: center;
 }
 
 .container {
