+++ ui/src/app/components/privacy-policy/privacy-policy.component.html
@@ -1,5 +1,5 @@
 <div fxLayout="column" fxLayoutAlign="center center" fxLayoutGap="5px" class="wrapper">
-    <span fxLayout="row" fxLayoutAlign="center center" fxLayoutGap="5px" class="title app-name" fxFlex="10">
+    <span fxLayout="row" fxLayoutAlign="center center" fxLayoutGap="5px" class="title page-title" fxFlex="10">
       Política de privacidad
     </span>
     <div fxLayout="column" fxLayoutAlign="start center" fxFlex="70">
