+++ ui/src/app/components/privacy-policy/privacy-policy.component.scss
@@ -2,10 +2,6 @@ mat-card {
   max-width: 80%;
 }
 
-.app-name {
-  font-size: 20px;
-}
-
 .info {
   p {
     letter-spacing: 0.1em;
