+++ ui/src/styles.scss
@@ -154,3 +154,8 @@ div.button {
     background: #fafafa;
   }
 }
+
+.page-title {
+  font-size: 22px;
+  text-align: center;
+}
